# sandbox/docs/src/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/run_code":{"post":{"tags":["sandbox","sandbox"],"summary":"Run Code","operationId":"run_code_run_code_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunCodeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunCodeResponse-Output"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/run_jupyter":{"post":{"tags":["sandbox","sandbox"],"summary":"Run Code In Jupyter","operationId":"Run_Code_in_Jupyter_run_jupyter_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunJupyterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunJupyterResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/list_datasets":{"get":{"tags":["datasets","datasets"],"summary":"List Datasets","description":"List all registered datasets","operationId":"list_datasets_list_datasets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"type":"string"},"type":"array","title":"Response List Datasets List Datasets Get"}}}}}}},"/list_ids":{"post":{"tags":["datasets","datasets"],"summary":"List Ids","description":"List all ids of a dataset","operationId":"list_ids_list_ids_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPromptsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"anyOf":[{"type":"integer"},{"type":"string"}]},"type":"array","title":"Response List Ids List Ids Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_prompts":{"post":{"tags":["datasets","datasets"],"summary":"Get Prompt","description":"Get prompts of a dataset","operationId":"get_prompt_get_prompts_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPromptsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Prompt"},"type":"array","title":"Response Get Prompt Get Prompts Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_prompt_by_id":{"post":{"tags":["datasets","datasets"],"summary":"Get Prompt By Id","description":"Get a single prompt given id in a dataset","operationId":"get_prompt_by_id_get_prompt_by_id_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPromptByIdRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Prompt"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/submit":{"post":{"tags":["datasets","datasets"],"summary":"Submit","description":"Submit a single problem in a dataset","operationId":"submit_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvalResult-Output"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_metrics":{"post":{"tags":["datasets","datasets"],"summary":"Get Metrics","description":"Get the metrics given all problem results in a dataset (partially supported)","operationId":"get_metrics_get_metrics_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetMetricsRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Metrics Get Metrics Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/get_metrics_function":{"post":{"tags":["datasets","datasets"],"summary":"Get Metrics Function","description":"Get the function to generate the metrics given results (partially supported)","operationId":"get_metrics_function_get_metrics_function_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetMetricsFunctionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetMetricsFunctionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ping":{"get":{"summary":"Index","operationId":"index_v1_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CellRunResult":{"properties":{"stdout":{"type":"string","title":"Stdout"},"stderr":{"type":"string","title":"Stderr"},"display":{"items":{"type":"object"},"type":"array","title":"Display"},"error":{"items":{"type":"object"},"type":"array","title":"Error"}},"type":"object","required":["stdout","stderr","display","error"],"title":"CellRunResult"},"CommandRunResult":{"properties":{"status":{"$ref":"#/components/schemas/CommandRunStatus"},"execution_time":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Execution Time"},"return_code":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Return Code"},"stdout":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stdout"},"stderr":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stderr"}},"type":"object","required":["status"],"title":"CommandRunResult"},"CommandRunStatus":{"type":"string","enum":["Finished","Error","TimeLimitExceeded"],"title":"CommandRunStatus"},"EvalResult-Input":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"string"}],"title":"Id"},"accepted":{"type":"boolean","title":"Accepted"},"extracted_code":{"type":"string","title":"Extracted Code"},"full_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Code"},"test_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Test Code"},"tests":{"items":{"$ref":"#/components/schemas/EvalTestCase-Input"},"type":"array","title":"Tests"},"extracted_type":{"anyOf":[{"type":"string","enum":["fenced","incomplete_fenced","heuristic","empty"]},{"type":"null"}],"title":"Extracted Type"},"extra":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Extra"}},"type":"object","required":["id","accepted","extracted_code","tests"],"title":"EvalResult"},"EvalResult-Output":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"string"}],"title":"Id"},"accepted":{"type":"boolean","title":"Accepted"},"extracted_code":{"type":"string","title":"Extracted Code"},"full_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Code"},"test_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Test Code"},"tests":{"items":{"$ref":"#/components/schemas/EvalTestCase-Output"},"type":"array","title":"Tests"},"extracted_type":{"anyOf":[{"type":"string","enum":["fenced","incomplete_fenced","heuristic","empty"]},{"type":"null"}],"title":"Extracted Type"},"extra":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Extra"}},"type":"object","required":["id","accepted","extracted_code","tests"],"title":"EvalResult"},"EvalTestCase-Input":{"properties":{"passed":{"type":"boolean","title":"Passed"},"exec_info":{"$ref":"#/components/schemas/RunCodeResponse-Input"},"test_info":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Test Info"}},"type":"object","required":["passed","exec_info"],"title":"EvalTestCase"},"EvalTestCase-Output":{"properties":{"passed":{"type":"boolean","title":"Passed"},"exec_info":{"$ref":"#/components/schemas/RunCodeResponse-Output"},"test_info":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Test Info"}},"type":"object","required":["passed","exec_info"],"title":"EvalTestCase"},"GetMetricsFunctionRequest":{"properties":{"dataset":{"type":"string","title":"Dataset"},"config":{"$ref":"#/components/schemas/TestConfig"}},"type":"object","required":["dataset","config"],"title":"GetMetricsFunctionRequest"},"GetMetricsFunctionResult":{"properties":{"function":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Function"}},"type":"object","required":["function"],"title":"GetMetricsFunctionResult"},"GetMetricsRequest":{"properties":{"dataset":{"type":"string","title":"Dataset"},"config":{"$ref":"#/components/schemas/TestConfig"},"results":{"items":{"$ref":"#/components/schemas/EvalResult-Input"},"type":"array","title":"Results"}},"type":"object","required":["dataset","config","results"],"title":"GetMetricsRequest"},"GetPromptByIdRequest":{"properties":{"dataset":{"type":"string","title":"Dataset"},"config":{"$ref":"#/components/schemas/TestConfig"},"id":{"anyOf":[{"type":"integer"},{"type":"string"}],"title":"Id"}},"type":"object","required":["dataset","config","id"],"title":"GetPromptByIdRequest"},"GetPromptsRequest":{"properties":{"dataset":{"type":"string","title":"Dataset"},"config":{"$ref":"#/components/schemas/TestConfig"},"offset":{"type":"integer","title":"Offset","default":0},"limit":{"type":"integer","title":"Limit","default":1000000}},"type":"object","required":["dataset","config"],"title":"GetPromptsRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Prompt":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"string"}],"title":"Id"},"prompt":{"type":"string","title":"Prompt"},"labels":{"type":"object","title":"Labels","default":{}}},"type":"object","required":["id","prompt"],"title":"Prompt"},"RunCodeRequest":{"properties":{"compile_timeout":{"type":"number","title":"Compile Timeout","description":"compile timeout for compiled languages","default":10},"run_timeout":{"type":"number","title":"Run Timeout","description":"code run timeout","default":10},"code":{"type":"string","title":"Code","description":"the code to run","examples":["print(\"hello\")"]},"stdin":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stdin","description":"optional string to pass into stdin","examples":[""]},"language":{"type":"string","enum":["python","cpp","nodejs","go","go_test","java","php","csharp","bash","typescript","sql","rust","cuda","lua","R","perl","D_ut","ruby","scala","julia","pytest","junit","kotlin_script","jest","verilog","python_gpu","lean","swift","racket"],"title":"Language","description":"the language or execution mode to run the code","examples":["python"]},"files":{"additionalProperties":{"type":"string"},"type":"object","title":"Files","description":"a dict from file path to base64 encoded file content","default":{}},"fetch_files":{"items":{"type":"string"},"type":"array","title":"Fetch Files","description":"a list of file paths to fetch after code execution","default":[]}},"type":"object","required":["code","language"],"title":"RunCodeRequest"},"RunCodeResponse-Input":{"properties":{"status":{"$ref":"#/components/schemas/RunStatus"},"message":{"type":"string","title":"Message"},"compile_result":{"anyOf":[{"$ref":"#/components/schemas/CommandRunResult"},{"type":"null"}]},"run_result":{"anyOf":[{"$ref":"#/components/schemas/CommandRunResult"},{"type":"null"}]},"executor_pod_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Executor Pod Name"},"files":{"additionalProperties":{"type":"string"},"type":"object","title":"Files","default":{}}},"type":"object","required":["status","message"],"title":"RunCodeResponse"},"RunCodeResponse-Output":{"properties":{"status":{"$ref":"#/components/schemas/RunStatus"},"message":{"type":"string","title":"Message"},"compile_result":{"anyOf":[{"$ref":"#/components/schemas/CommandRunResult"},{"type":"null"}]},"run_result":{"anyOf":[{"$ref":"#/components/schemas/CommandRunResult"},{"type":"null"}]},"executor_pod_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Executor Pod Name"},"files":{"additionalProperties":{"type":"string"},"type":"object","title":"Files","default":{}}},"type":"object","required":["status","message"],"title":"RunCodeResponse"},"RunJupyterRequest":{"properties":{"cells":{"items":{"type":"string"},"type":"array","title":"Cells","description":"list of code blocks to run in jupyter notebook","examples":[["a = 123","a","print(a)","import sys; sys.stderr.write(\"stderr message\")","raise RuntimeError(\"error message\")"]]},"cell_timeout":{"type":"number","title":"Cell Timeout","description":"max run time for each of the cells","default":0},"total_timeout":{"type":"number","title":"Total Timeout","description":"max run time for all the cells","default":45},"kernel":{"const":"python3","title":"Kernel","default":"python3"},"files":{"additionalProperties":{"type":"string"},"type":"object","title":"Files","description":"a dict from file path to base64 encoded file content","default":{}},"fetch_files":{"items":{"type":"string"},"type":"array","title":"Fetch Files","description":"a list of file paths to fetch after code execution","default":[]}},"type":"object","required":["cells"],"title":"RunJupyterRequest"},"RunJupyterResponse":{"properties":{"status":{"$ref":"#/components/schemas/RunStatus"},"message":{"type":"string","title":"Message"},"driver":{"anyOf":[{"$ref":"#/components/schemas/CommandRunResult"},{"type":"null"}]},"cells":{"items":{"$ref":"#/components/schemas/CellRunResult"},"type":"array","title":"Cells","default":[]},"executor_pod_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Executor Pod Name"},"files":{"additionalProperties":{"type":"string"},"type":"object","title":"Files","default":{}}},"type":"object","required":["status","message"],"title":"RunJupyterResponse"},"RunStatus":{"type":"string","enum":["Success","Failed","SandboxError"],"title":"RunStatus"},"SubmitRequest":{"properties":{"dataset":{"type":"string","title":"Dataset"},"id":{"anyOf":[{"type":"integer"},{"type":"string"}],"title":"Id"},"completion":{"type":"string","title":"Completion"},"config":{"$ref":"#/components/schemas/TestConfig"}},"type":"object","required":["dataset","id","completion","config"],"title":"SubmitRequest"},"TestConfig":{"properties":{"dataset_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Type","description":"the dataset class used to process, only works when the dataset id is not registered.","examples":[null]},"language":{"anyOf":[{"type":"string","enum":["python","cpp","nodejs","go","go_test","java","php","csharp","bash","typescript","sql","rust","cuda","lua","R","perl","D_ut","ruby","scala","julia","pytest","junit","kotlin_script","jest","verilog","python_gpu","lean","swift","racket"]},{"type":"null"}],"title":"Language"},"locale":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Locale"},"is_fewshot":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Fewshot"},"compile_timeout":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Compile Timeout"},"run_timeout":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Run Timeout"},"custom_extract_logic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Custom Extract Logic"},"provided_data":{"anyOf":[{"items":{"type":"object"},"type":"array"},{"type":"object"},{"type":"null"}],"title":"Provided Data"},"extra":{"type":"object","title":"Extra","default":{}}},"type":"object","title":"TestConfig","description":"custom_extract_logic: a piece of python code that calls `submit_code_blocks(cbs)` to extract custom code\n                      cbs: List[CodeBlock], CodeBlock(priority=40, code='xxx', language='xxx')\n                      priority: fenced = 30, incomplete fenced = 20, heuristic = 10"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}